requirements.txt
Flask
bs4
mysql-connector
google.generativeai
flask_cors